feat: Mac MPS training support and progress robustness (v2.3.0)
- Force model to float32 on MPS to avoid bf16 NaN on Apple Silicon
- Compute mse_loss in fp32 to match autocast convention and fix MPS dtype mismatch
- Sanitize NaN/Inf in WebSocket payloads so progress events stay valid JSON
- Drive node progress bar at 100-tick resolution for smoother updates

+++ src/voxcpm/modules/locdit/unified_cfm.py
@@ -221,7 +221,7 @@ class UnifiedCFM(torch.nn.Module):
             u_pred = model_fn(y, r, t)
             u_tgt = v
 
-        losses = F.mse_loss(u_pred, u_tgt.detach(), reduction="none").mean(dim=1)
+        losses = F.mse_loss(u_pred.float(), u_tgt.detach().float(), reduction="none").mean(dim=1)
         if tgt_mask is not None:
             weights = self.adaptive_loss_weighting(losses, tgt_mask.squeeze(1))
             loss = (weights * losses).sum() / torch.sum(tgt_mask)
